+++ smac/examples/check_style.sh
@@ -1,0 +1,1 @@
+flake8 --ignore "E501,E731,E74,E402,F401,W503,F405,E121,E122,E126" pettingzoo/
