+++ smac/examples/pettingzoo/smac_env/sc2_v0.py
@@ -7,8 +7,10 @@ from gym.utils import EzPickle
 
 max_cycles_default = 1000
 
+
 def parallel_env(max_cycles=max_cycles_default, **smac_args):
     return _parallel_env(max_cycles, **smac_args)
+
 
 def raw_env(max_cycles=max_cycles_default, **smac_args):
     return from_parallel_wrapper(parallel_env(max_cycles, **smac_args))
@@ -18,10 +20,9 @@ env = make_env(raw_env)
 
 
 class _parallel_env(smac_parallel_env, EzPickle):
-    metadata = {'render.modes': ['human', 'rgb_array'], 'name': "sc2_v0"}
+    metadata = {'render.modes': ['human'], 'name': "sc2_v0"}
 
     def __init__(self, max_cycles, **smac_args):
         EzPickle.__init__(self, max_cycles, **smac_args)
         env = StarCraft2Env(**smac_args)
         super().__init__(env, max_cycles)
-
