+++ smac/examples/pettingzoo/smactest.py
@@ -2,20 +2,12 @@ import time
 import random
 import numpy as np
 
-from smac.examples.pettingzoo import sc2_v0
-
+from smac_env import sc2_v0
 
 def random_demo(env, render=True, cycles=100):
     '''
     Runs an env object with random actions.
     '''
-
-    env.reset()
-
-    if hasattr(env, 'display_wait'):
-        display_wait = env.display_wait
-    else:
-        display_wait = 0.0
 
     total_reward = 0
     done = False
@@ -23,13 +15,12 @@ def random_demo(env, render=True, cycles=100):
     for cycle in range(cycles):
         env.reset()
         for agent in env.agent_iter():
+
             if render:
                 env.render()
-                time.sleep(display_wait)
 
             obs, reward, done, info = env.last()
             total_reward += reward
-            
             if done:
                 action = None
             elif isinstance(obs, dict) and 'action_mask' in obs:
@@ -37,6 +28,7 @@ def random_demo(env, render=True, cycles=100):
             else:
                 action = env.action_spaces[agent].sample()
             env.step(action)
+
     print("Total reward", total_reward, "done", done)
 
     if render:
@@ -47,10 +39,8 @@ def random_demo(env, render=True, cycles=100):
 
     return total_reward
 
-
-
 def main():
-    env = sc2_v0.env(map_name = "MMM", step_mul=2)
+    env = sc2_v0.env(map_name = "25m", step_mul=2, window_size_x=1000, window_size_y=1000)
     random_demo(env)
 
 if __name__ == "__main__":
